This is the intro text that pops up
This is the intro text that uses an basic array of tezt to print on the screen asap. This can of course create multiple bugs, since it has no real idea of when the player actually sees it. Also, if there is a syntex error, the entire script will stop. Also it never gets deleted, thus using ram on the server machine creating even more fps problems for the users.